requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Flask-WTF==1.1.1
python-dotenv==1.0.0
gunicorn==21.2.0
Jinja2==3.1.2
